+++ src/components/baan-selection/ListBaan.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import { convertSize, isAdded } from '../utils/utility';
+import { convertSize, isAdded } from '@/utils/baan-selection/utility';
 import { IBaan } from '@/types/baan';
 const ListBaan = (props: any) => {
     return props.baan
